:truck::art:restructured backend/API to be object oriented.
