+++ serverless.ts
@@ -1,6 +1,6 @@
 import type { AWS } from '@serverless/typescript';
 import { getAllCreators, addCreator, loginCreator } from '@functions/creator';
-import { search } from '@functions/search';
+import { search } from '@functions/tweet';
 import { CreatorTable } from '@model/creator/index';
 import { ResultSetTable } from '@model/resultSet';
 import { TweetTable } from '@model/tweet';
@@ -10,7 +10,7 @@ import { TextStylesTable } from '@model/textStyles';
 import { getAllResultSet, getResultSet } from '@functions/resultSet';
 import { generateReport } from '@functions/generateReport';
 import { getAllMyReports, getAllReports, getReport } from '@functions/report';
-import { editBlock } from '@functions/editBlock';
+import { editBlock } from '@functions/reportBlock';
 
 
 const serverlessConfiguration: AWS = {
