+++ src/functions/report/index.ts
@@ -41,3 +41,17 @@ export const getReport = {
         },
     ],
 };
+
+export const cloneReport = {
+    handler: `${handlerPath(__dirname)}/handler.cloneReport`,
+    description: 'A function that clones a report.',
+    events: [
+        {
+            http: {
+                method: 'get',
+                path: 'cloneReport/',
+                cors: true
+            },
+        },
+    ],
+};
