--- src/functions/trending/handler.ts
@@ -1,4 +1,0 @@
-import { APIGatewayProxyEvent, APIGatewayProxyResult } from "aws-lambda";
-import { formatJSONResponse } from '@libs/api-gateway';
-import { middyfy } from '@libs/lambda';
-import CreatorServices from "../../services";
